improve: commonize code
