+++ src/comments/BaseComment.ts
@@ -3,7 +3,12 @@ import type {
   formattedCommentWithSize,
   IComment,
 } from "@/@types/";
-import { config } from "@/definition/config";
+import {
+  formattedCommentWithFont,
+  measureTextInput,
+  measureTextResult,
+} from "@/@types/";
+import { config, options } from "@/definition/config";
 import { NotImplementedError } from "@/errors/";
 import { getPosX, isBanActive, isReverseActive, parseFont } from "@/utils";
 
@@ -48,6 +53,54 @@ class BaseComment implements IComment {
   }
   get mail() {
     return this.comment.mail;
+  }
+
+  /**
+   * コメントの描画サイズを計算する
+   */
+  getCommentSize(
+    parsedData: formattedCommentWithFont
+  ): formattedCommentWithSize {
+    this.context.font = parseFont(parsedData.font, parsedData.fontSize);
+    const size = parsedData as formattedCommentWithSize;
+    if (parsedData.invisible) {
+      size.height = 0;
+      size.width = 0;
+      size.lineHeight = 0;
+      size.fontSize = 0;
+      size.content = [];
+      size.resized = false;
+      size.resizedX = false;
+      size.resizedY = false;
+      size.charSize = 0;
+      return size;
+    }
+    const measure = this.measureText(parsedData);
+    if (options.scale !== 1 && size.layer === -1) {
+      measure.height *= options.scale;
+      measure.width *= options.scale;
+      measure.fontSize *= options.scale;
+    }
+    size.height = measure.height;
+    size.width = measure.width;
+    size.lineHeight = measure.lineHeight;
+    size.fontSize = measure.fontSize;
+    size.content = measure.content;
+    size.resized = measure.resized;
+    size.resizedX = measure.resizedX;
+    size.resizedY = measure.resizedY;
+    size.charSize = measure.charSize;
+    return size;
+  }
+  /**
+   * context.measureTextの複数行対応版
+   * 画面外にはみ出すコメントの縮小も行う
+   * @param comment - 独自フォーマットのコメントデータ
+   * @returns {{resized: boolean, width: number, width: number, fontSize: number, width_min: number, height: number, lineHeight: number}} - 描画サイズとリサイズの情報
+   */
+  measureText(comment: measureTextInput): measureTextResult {
+    console.error("measureText method is not implemented", comment);
+    throw new NotImplementedError(this.pluginName, "measureText");
   }
   convertComment(comment: formattedComment): formattedCommentWithSize {
     console.error("convertComment method is not implemented", comment);
