+++ src/comments/HTML5Comment.ts
@@ -67,7 +67,7 @@ class HTML5Comment extends BaseComment {
    * @param comment - 独自フォーマットのコメントデータ
    * @returns {{resized: boolean, width: number, width: number, fontSize: number, width_min: number, height: number, lineHeight: number}} - 描画サイズとリサイズの情報
    */
-  measureText(comment: measureTextInput): measureTextResult {
+  override measureText(comment: measureTextInput): measureTextResult {
     const widthLimit = getConfig(config.commentStageSize, false)[
         comment.full ? "fullWidth" : "width"
       ],
@@ -152,44 +152,6 @@ class HTML5Comment extends BaseComment {
       resizedY: !!comment.resizedY,
       charSize: comment.charSize || 0,
     };
-  }
-
-  /**
-   * コメントの描画サイズを計算する
-   */
-  getCommentSize(
-    parsedData: formattedCommentWithFont
-  ): formattedCommentWithSize {
-    this.context.font = parseFont(parsedData.font, parsedData.fontSize);
-    const size = parsedData as formattedCommentWithSize;
-    if (parsedData.invisible) {
-      size.height = 0;
-      size.width = 0;
-      size.lineHeight = 0;
-      size.fontSize = 0;
-      size.content = [];
-      size.resized = false;
-      size.resizedX = false;
-      size.resizedY = false;
-      size.charSize = 0;
-      return size;
-    }
-    const measure = this.measureText(parsedData);
-    if (options.scale !== 1 && size.layer === -1) {
-      measure.height *= options.scale;
-      measure.width *= options.scale;
-      measure.fontSize *= options.scale;
-    }
-    size.height = measure.height;
-    size.width = measure.width;
-    size.lineHeight = measure.lineHeight;
-    size.fontSize = measure.fontSize;
-    size.content = measure.content;
-    size.resized = measure.resized;
-    size.resizedX = measure.resizedX;
-    size.resizedY = measure.resizedY;
-    size.charSize = measure.charSize;
-    return size;
   }
 
   override _drawCollision(posX: number, posY: number, showCollision: boolean) {
